+++ docs/source/conf.py
@@ -1,3 +1,34 @@
+"""
+Notes:
+    http://docs.readthedocs.io/en/latest/getting_started.html
+
+    pip install sphinx sphinx-autobuild sphinx_rtd_theme sphinxcontrib-napoleon
+
+    cd ~/code/timerit
+    mkdir docs
+    cd docs
+
+    sphinx-quickstart
+
+    # need to edit the conf.py
+
+    cd ~/code/timerit/docs
+    make html
+    sphinx-apidoc -f -o ~/code/timerit/docs/source ~/code/timerit/timerit --separate
+    make html
+
+
+    Also:
+        To turn on PR checks
+
+        https://docs.readthedocs.io/en/stable/guides/autobuild-docs-for-pull-requests.html
+
+        https://readthedocs.org/dashboard/timerit/advanced/
+
+        ensure your github account is connected to readthedocs
+        https://readthedocs.org/accounts/social/connections/
+
+"""
 # -*- coding: utf-8 -*-
 #
 # Configuration file for the Sphinx documentation builder.
@@ -12,24 +43,42 @@
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
-import os
-import sys
-sys.path.insert(0, os.path.abspath('.'))
-sys.path.insert(0, os.path.abspath('..'))
-sys.path.insert(0, os.path.abspath('../..'))
+import sphinx_rtd_theme
+from os.path import exists
+from os.path import dirname
+from os.path import join
+
+
+def parse_version(fpath):
+    """
+    Statically parse the version number from a python file
+    """
+    import ast
+    if not exists(fpath):
+        raise ValueError('fpath={!r} does not exist'.format(fpath))
+    with open(fpath, 'r') as file_:
+        sourcecode = file_.read()
+    pt = ast.parse(sourcecode)
+    class VersionVisitor(ast.NodeVisitor):
+        def visit_Assign(self, node):
+            for target in node.targets:
+                if getattr(target, 'id', None) == '__version__':
+                    self.version = node.value.s
+    visitor = VersionVisitor()
+    visitor.visit(pt)
+    return visitor.version
 
 
 # -- Project information -----------------------------------------------------
 
-project = 'timerit'
+project = 'TimerIt'
 copyright = '2019, Jon Crall'
 author = 'Jon Crall'
+modname = 'timerit'
 
-# The short X.Y version
-import timerit  # NOQA
-version = '.'.join(timerit.__version__.split('.')[0:2])
-# The full version, including alpha/beta/rc tags
-release = ''
+modpath = join(dirname(dirname(dirname(__file__))), modname, '__init__.py')
+release = parse_version(modpath)
+version = '.'.join(release.split('.')[0:2])
 
 
 # -- General configuration ---------------------------------------------------
@@ -183,8 +232,140 @@ napoleon_use_ivar = True
 autodoc_inherit_docstrings = False
 autodoc_member_order = 'bysource'
 
+# autodoc_mock_imports = ['torch', 'torchvision', 'visdom']
+intersphinx_mapping = {
+    # 'pytorch': ('http://pytorch.org/docs/master/', None),
+    'python': ('https://docs.python.org/3', None),
+    'click': ('https://click.palletsprojects.com/', None),
+    # 'xxhash': ('https://pypi.org/project/xxhash/', None),
+    # 'pygments': ('https://pygments.org/docs/', None),
+    # 'tqdm': ('https://tqdm.github.io/', None),
+}
+__dev_note__ = """
+python -m sphinx.ext.intersphinx https://docs.python.org/3/objects.inv
+python -m sphinx.ext.intersphinx https://ubelt.readthedocs.io/en/latest/objects.inv
+python -m sphinx.ext.intersphinx https://networkx.org/documentation/stable/objects.inv
+"""
+
+
 html_theme_options = {
     'collapse_navigation': False,
     'display_version': True,
     # 'logo_only': True,
 }
+
+
+# -- Extension configuration -------------------------------------------------
+
+
+from sphinx.domains.python import PythonDomain  # NOQA
+# from sphinx.application import Sphinx  # NOQA
+from typing import Any, List  # NOQA
+
+
+class PatchedPythonDomain(PythonDomain):
+    """
+    References:
+        https://github.com/sphinx-doc/sphinx/issues/3866
+    """
+    def resolve_xref(self, env, fromdocname, builder, typ, target, node, contnode):
+        # TODO: can use this to resolve references nicely
+        if target.startswith('ub.'):
+            target = 'ubelt.' + target[3]
+        return_value = super(PatchedPythonDomain, self).resolve_xref(
+            env, fromdocname, builder, typ, target, node, contnode)
+        return return_value
+
+
+def process(app, what_: str, name: str, obj: Any, options: Any, lines:
+            List[str]) -> None:
+    """
+    Custom process to transform docstring lines Remove "Ignore" blocks
+
+    Args:
+        app (sphinx.application.Sphinx): the Sphinx application object
+
+        what (str):
+            the type of the object which the docstring belongs to (one of
+            "module", "class", "exception", "function", "method", "attribute")
+
+        name (str): the fully qualified name of the object
+
+        obj: the object itself
+
+        options: the options given to the directive: an object with
+            attributes inherited_members, undoc_members, show_inheritance
+            and noindex that are true if the flag option of same name was
+            given to the auto directive
+
+        lines (List[str]): the lines of the docstring, see above
+
+    References:
+        https://www.sphinx-doc.org/en/1.5.1/_modules/sphinx/ext/autodoc.html
+        https://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html
+    """
+    # if what and what_ not in what:
+    #     return
+    orig_lines = lines[:]
+
+    # text = '\n'.join(lines)
+    # if 'Example' in text and 'CommandLine' in text:
+    #     import xdev
+    #     xdev.embed()
+
+    ignore_tags = tuple(['Ignore'])
+
+    mode = None
+    # buffer = None
+    new_lines = []
+    for i, line in enumerate(orig_lines):
+
+        # See if the line triggers a mode change
+        if line.startswith(ignore_tags):
+            mode = 'ignore'
+        elif line.startswith('CommandLine'):
+            mode = 'cmdline'
+        elif line and not line.startswith(' '):
+            # if the line startswith anything but a space, we are no
+            # longer in the previous nested scope
+            mode = None
+
+        if mode is None:
+            new_lines.append(line)
+        elif mode == 'ignore':
+            # print('IGNORE line = {!r}'.format(line))
+            pass
+        elif mode == 'cmdline':
+            if line.startswith('CommandLine'):
+                new_lines.append('.. rubric:: CommandLine')
+                new_lines.append('')
+                new_lines.append('.. code-block:: bash')
+                new_lines.append('')
+                # new_lines.append('    # CommandLine')
+            else:
+                # new_lines.append(line.strip())
+                new_lines.append(line)
+        else:
+            raise KeyError(mode)
+
+    lines[:] = new_lines
+    # make sure there is a blank line at the end
+    if lines and lines[-1]:
+        lines.append('')
+
+
+def setup(app):
+    app.add_domain(PatchedPythonDomain, override=True)
+    if 1:
+        # New Way
+        # what = None
+        app.connect('autodoc-process-docstring', process)
+    else:
+        # OLD WAY
+        # https://stackoverflow.com/questions/26534184/can-sphinx-ignore-certain-tags-in-python-docstrings
+        # Register a sphinx.ext.autodoc.between listener to ignore everything
+        # between lines that contain the word IGNORE
+        # from sphinx.ext.autodoc import between
+        # app.connect('autodoc-process-docstring', between('^ *Ignore:$', exclude=True))
+        pass
+    return app