+++ timerit/__init__.py
@@ -23,20 +23,20 @@ a single line, but including more is trivial.
     >>>     setup_vars = 10000
     >>>     with timer:
     >>>         math.factorial(setup_vars)
+    >>> # xdoctest: +IGNORE_WANT
     >>> print('t1.total_time = %r' % (t1.total_time,))
     Timing for 200 loops
     Timed for: 200 loops, best of 3
         time per loop: best=2.064 ms, mean=2.115 +- 0.05 ms
     t1.total_time = 0.4427177629695507
-
-
-Mkinit:
-    mkinit timerit --nomods --relative
-    # TODO: get sphinx to ignore these mkinit-blocks
 """
-from __future__ import unicode_literals
+__version__ = '1.0.0'
 
-__version__ = '0.3.0'
+__mkinit__ = """
+# Autogen command
+mkinit timerit --nomods --relative
+"""
 
 from .core import (Timer, Timerit,)
+
 __all__ = ['Timer', 'Timerit']